Add adder functionality and refactor TypeScript integration
